# internal/api/oas/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package oas

import (
	"time"

	"github.com/go-faster/errors"
	"github.com/go-faster/jx"
	"github.com/google/uuid"
)

type CreateTaskBadRequest Error

func (*CreateTaskBadRequest) createTaskRes() {}

type CreateTaskInternalServerError Error

func (*CreateTaskInternalServerError) createTaskRes() {}

// Ref: #/components/schemas/CreateTaskRequest
type CreateTaskRequest struct {
	Handler        string                      `json:"handler"`
	Payload        OptCreateTaskRequestPayload `json:"payload"`
	IdempotencyKey OptString                   `json:"idempotency_key"`
}

// GetHandler returns the value of Handler.
func (s *CreateTaskRequest) GetHandler() string {
	return s.Handler
}

// GetPayload returns the value of Payload.
func (s *CreateTaskRequest) GetPayload() OptCreateTaskRequestPayload {
	return s.Payload
}

// GetIdempotencyKey returns the value of IdempotencyKey.
func (s *CreateTaskRequest) GetIdempotencyKey() OptString {
	return s.IdempotencyKey
}

// SetHandler sets the value of Handler.
func (s *CreateTaskRequest) SetHandler(val string) {
	s.Handler = val
}

// SetPayload sets the value of Payload.
func (s *CreateTaskRequest) SetPayload(val OptCreateTaskRequestPayload) {
	s.Payload = val
}

// SetIdempotencyKey sets the value of IdempotencyKey.
func (s *CreateTaskRequest) SetIdempotencyKey(val OptString) {
	s.IdempotencyKey = val
}

type CreateTaskRequestPayload map[string]jx.Raw

func (s *CreateTaskRequestPayload) init() CreateTaskRequestPayload {
	m := *s
	if m == nil {
		m = map[string]jx.Raw{}
		*s = m
	}
	return m
}

// Ref: #/components/schemas/Error
type Error struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// GetCode returns the value of Code.
func (s *Error) GetCode() string {
	return s.Code
}

// GetMessage returns the value of Message.
func (s *Error) GetMessage() string {
	return s.Message
}

// SetCode sets the value of Code.
func (s *Error) SetCode(val string) {
	s.Code = val
}

// SetMessage sets the value of Message.
func (s *Error) SetMessage(val string) {
	s.Message = val
}

func (*Error) getTaskRes() {}

// NewOptCreateTaskRequestPayload returns new OptCreateTaskRequestPayload with value set to v.
func NewOptCreateTaskRequestPayload(v CreateTaskRequestPayload) OptCreateTaskRequestPayload {
	return OptCreateTaskRequestPayload{
		Value: v,
		Set:   true,
	}
}

// OptCreateTaskRequestPayload is optional CreateTaskRequestPayload.
type OptCreateTaskRequestPayload struct {
	Value CreateTaskRequestPayload
	Set   bool
}

// IsSet returns true if OptCreateTaskRequestPayload was set.
func (o OptCreateTaskRequestPayload) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptCreateTaskRequestPayload) Reset() {
	var v CreateTaskRequestPayload
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptCreateTaskRequestPayload) SetTo(v CreateTaskRequestPayload) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptCreateTaskRequestPayload) Get() (v CreateTaskRequestPayload, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptCreateTaskRequestPayload) Or(d CreateTaskRequestPayload) CreateTaskRequestPayload {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// NewOptString returns new OptString with value set to v.
func NewOptString(v string) OptString {
	return OptString{
		Value: v,
		Set:   true,
	}
}

// OptString is optional string.
type OptString struct {
	Value string
	Set   bool
}

// IsSet returns true if OptString was set.
func (o OptString) IsSet() bool { return o.Set }

// Reset unsets value.
func (o *OptString) Reset() {
	var v string
	o.Value = v
	o.Set = false
}

// SetTo sets value to v.
func (o *OptString) SetTo(v string) {
	o.Set = true
	o.Value = v
}

// Get returns value and boolean that denotes whether value was set.
func (o OptString) Get() (v string, ok bool) {
	if !o.Set {
		return v, false
	}
	return o.Value, true
}

// Or returns value if set, or given parameter if does not.
func (o OptString) Or(d string) string {
	if v, ok := o.Get(); ok {
		return v
	}
	return d
}

// Ref: #/components/schemas/Task
type Task struct {
	ID        uuid.UUID  `json:"id"`
	Status    TaskStatus `json:"status"`
	CreatedAt time.Time  `json:"created_at"`
}

// GetID returns the value of ID.
func (s *Task) GetID() uuid.UUID {
	return s.ID
}

// GetStatus returns the value of Status.
func (s *Task) GetStatus() TaskStatus {
	return s.Status
}

// GetCreatedAt returns the value of CreatedAt.
func (s *Task) GetCreatedAt() time.Time {
	return s.CreatedAt
}

// SetID sets the value of ID.
func (s *Task) SetID(val uuid.UUID) {
	s.ID = val
}

// SetStatus sets the value of Status.
func (s *Task) SetStatus(val TaskStatus) {
	s.Status = val
}

// SetCreatedAt sets the value of CreatedAt.
func (s *Task) SetCreatedAt(val time.Time) {
	s.CreatedAt = val
}

func (*Task) createTaskRes() {}
func (*Task) getTaskRes()    {}

type TaskStatus string

const (
	TaskStatusPending   TaskStatus = "pending"
	TaskStatusRunning   TaskStatus = "running"
	TaskStatusSucceeded TaskStatus = "succeeded"
	TaskStatusFailed    TaskStatus = "failed"
	TaskStatusCancelled TaskStatus = "cancelled"
)

// AllValues returns all TaskStatus values.
func (TaskStatus) AllValues() []TaskStatus {
	return []TaskStatus{
		TaskStatusPending,
		TaskStatusRunning,
		TaskStatusSucceeded,
		TaskStatusFailed,
		TaskStatusCancelled,
	}
}

// MarshalText implements encoding.TextMarshaler.
func (s TaskStatus) MarshalText() ([]byte, error) {
	switch s {
	case TaskStatusPending:
		return []byte(s), nil
	case TaskStatusRunning:
		return []byte(s), nil
	case TaskStatusSucceeded:
		return []byte(s), nil
	case TaskStatusFailed:
		return []byte(s), nil
	case TaskStatusCancelled:
		return []byte(s), nil
	default:
		return nil, errors.Errorf("invalid value: %q", s)
	}
}

// UnmarshalText implements encoding.TextUnmarshaler.
func (s *TaskStatus) UnmarshalText(data []byte) error {
	switch TaskStatus(data) {
	case TaskStatusPending:
		*s = TaskStatusPending
		return nil
	case TaskStatusRunning:
		*s = TaskStatusRunning
		return nil
	case TaskStatusSucceeded:
		*s = TaskStatusSucceeded
		return nil
	case TaskStatusFailed:
		*s = TaskStatusFailed
		return nil
	case TaskStatusCancelled:
		*s = TaskStatusCancelled
		return nil
	default:
		return errors.Errorf("invalid value: %q", data)
	}
}